Changing vertex_times type from int to float

+++ NuRadioMC/EvtGen/generator.py
@@ -897,7 +897,7 @@ def generate_surface_muons(filename, n_events, Emin, Emax,
 
     data_sets["event_ids"] = np.arange(n_events) + start_event_id
     data_sets["n_interaction"] = np.ones(n_events, dtype=np.int)
-    data_sets["vertex_times"] = np.zeros(n_events, dtype=np.int)
+    data_sets["vertex_times"] = np.zeros(n_events, dtype=np.float)
 
     # generate neutrino flavors randomly
 
@@ -1228,7 +1228,7 @@ def generate_eventlist_cylinder(filename, n_events, Emin, Emax,
     data_sets["event_ids"] = np.arange(n_events) + start_event_id
     logger.debug("generating number of interactions")
     data_sets["n_interaction"] = np.ones(n_events, dtype=np.int)
-    data_sets["vertex_times"] = np.zeros(n_events, dtype=np.int)
+    data_sets["vertex_times"] = np.zeros(n_events, dtype=np.float)
 
     # generate neutrino flavors randomly
     logger.debug("generating flavors")
